requirements.txt
ipython==8.2.0
jupyter==1.0.0
matplotlib==3.5.1
numpy==1.22.3
torch==1.11.0
torchvision==0.12.0
pandas
Pillow
gdown